fix: resolve static-analysis findings (zero warnings/errors)
Production correctness fixes:
- App: replace atoi (no error reporting) with a from_chars-based clamped parse;
  merge the duplicate QUIT/ESCAPE event branches.
- IblProbe/ProceduralGeometry/test: widen before multiplying in index/reserve
  arithmetic (avoid int overflow before size_t conversion).
- Texture: explicit float cast when normalising bytes; use std::lround instead
  of (x+0.5) rounding; fix int->float narrowing.
- Logger: remove dead store (default branch now assigns the UNKNOWN fallback);
  std::lock_guard -> std::scoped_lock.
- SceneBase: drop no-op std::move on a trivially-copyable Material.
- SceneLoader: value-initialise LoadedSubmesh::meshIndex (cppcheck).
- Context::create: take Config by const ref (only read, never moved).
- AccumulationPass/SceneOutputCopyPass: mark destructors override.
- Timer: NOLINT the intentional empty catch in the noexcept destructor.

Profile refinement (canonical):
- .clang-tidy: disable use-nodiscard, use-std-numbers, use-auto, use-ranges,
  avoid-c-style-cast, performance-enum-size, bugprone-invalid-enum-default-
  initialization (noisy/subjective/pervasive-in-Vulkan).
- tests/.clang-tidy: mirror those + relax bugprone-random-generator-seed
  (deterministic seeds are intentional in tests).
- CMakeLists: exclude the stb/VMA third-party impl-wrapper TUs from clang-tidy.
- Convert the remaining tests/ #pragma once to include guards.

clang-tidy + cppcheck now emit zero warnings and zero errors; 73 tests green.

+++ src/harmonia/scene/ProceduralGeometry.cpp
@@ -70,8 +70,8 @@ MeshData makeBox(sm::float3 halfExtent, const sm::float4x4& transform) {
 
 MeshData makeSphere(sm::float3 center, float radius, uint32_t rings, uint32_t slices) {
     MeshData mesh;
-    mesh.vertices.reserve((rings + 1) * (slices + 1));
-    mesh.indices.reserve(rings * slices * 6);
+    mesh.vertices.reserve(static_cast<size_t>(rings + 1) * static_cast<size_t>(slices + 1));
+    mesh.indices.reserve(static_cast<size_t>(rings) * static_cast<size_t>(slices) * 6u);
 
     const float pi = sm::pi<float>();
